Add HNSW vector store, and update the embedding format (#507)
- Add a registry for vector stores, analogous to signals and sources
- Set the default vector store to `hnsw`.
- Let vector stores write their own index to disk and read from disk
(since index creation can take a while)
- Fix the placeholders for the LHS menu. Server slowed down on purpose:


https://github.com/lilacai/lilac/assets/2294279/53a76eed-3e2a-47fe-8417-d035b4f063cd

+++ lilac/embeddings/embedding.py
@@ -87,8 +87,7 @@ def compute_split_embeddings(docs: Iterable[str],
 
     for x in list(pool.map(lambda x: embed_fn(x), chunks(texts, batch_size))):
       embeddings.extend(x)
-    matrix = cast(np.ndarray, normalize(np.array(embeddings)))
-    matrix = matrix.astype(np.float16)
+    matrix = cast(np.ndarray, normalize(np.array(embeddings, dtype=np.float32)))
     # np.split returns a shallow copy of each embedding so we don't increase the mem footprint.
     embeddings_batch = cast(list[np.ndarray], np.split(matrix, matrix.shape[0]))
     for (index, (_, (start, end))), embedding in zip(batch, embeddings_batch):
